+++ sensirion_uart_scc1/drivers/scc1_ld20.py
@@ -1,19 +1,18 @@
 # -*- coding: utf-8 -*-
 
-from sensirion_uart_scc1.drivers.scc1_sf06 import Scc1Sf06
 from sensirion_uart_scc1.drivers.slf_common import SlfMode, SLF_PRODUCT_LIQUI_MAP, SlfProduct
 from sensirion_uart_scc1.scc1_shdlc_device import Scc1ShdlcDevice
+from sensirion_uart_scc1.drivers.scc1_sf06 import Scc1Sf06
 
 
-class Scc1Slf3x(Scc1Sf06):
+class Scc1Ld20(Scc1Sf06):
     """
-    Scc1 SLF3x Sensor Driver
+    Scc1 LD20 Sensor Driver
 
-    The Scc1 provides features to support the SLF3x liquid flow sensors. This driver accesses the sensor through the
+    The Scc1 provides features to support the LD20 liquid flow sensors. This driver accesses the sensor through the
     API specified by scc1.
     """
-    SENSOR_TYPE = 3  #: Sensor type for SLF3x
-    START_MEASUREMENT_DELAY_S = 0.015
+    SENSOR_TYPE = 3  #: Sensor type for LD20 (same as SF06/SLF3x)
 
     def __init__(self, device: Scc1ShdlcDevice, liquid_mode: SlfMode = SlfMode.LIQUI_1) -> None:
         """
@@ -23,4 +22,4 @@ class Scc1Slf3x(Scc1Sf06):
         :param liquid_mode: The liquid that is measured.
         """
         super().__init__(device, liquid_mode)
-        self._liquid_config = SLF_PRODUCT_LIQUI_MAP[SlfProduct.SLF3x]
+        self._liquid_config = SLF_PRODUCT_LIQUI_MAP[SlfProduct.LD20]